# openapi/mocknetpack.yaml
openapi: 3.0.3
info:
  title: MockNetPack API
  version: 0.13.0
  description: |
    MockNetPack 中央 Mock 服务器 API 契约。
    - AI 接入层：为 Agent 工具与脚本/CLI 提供机器原生通道——
      (1) 长效 API Key（POST/GET /api/v1/auth/keys、DELETE /api/v1/auth/keys/{id}）：创建时明文仅返回一次，
      服务端只存 SHA-256 哈希；`Authorization: Bearer <apiKey>` 与登录会话 token 并存，鉴权语义一致；
      (2) 流量数据体验补强：GET /sessions/{id}/traffic 新增 projection=compact（8 字段瘦身视图 +
      每会话自增 seq 增量游标）、since（seq 增量拉取）、method/scheme 过滤，默认（非 compact）响应形状不变，
      人类 Web UI 体验不回归；(3) 错误响应按「给 AI 的指引」：401 提示配置 API Key、404/409/403 说明原因与下一步。
    - 权限与分享（/，v0.11.0）：MockRule 新增 owner（创建者用户名，服务端从当前会话用户写入，
      创建时设）与 updatedBy（最后修改者，每次 PUT 更新）两个纯服务端字段——Web 全量列表下发展示；
      SDK 增量拉取（sinceVersion）只返回 SDK 消费字段（id/method/path/response/enabled/effective），
      绝不携带 owner/updatedBy 等纯服务端/Web-only 字段。权限矩阵：查看规则对设备可见者全部开放
      （设备隔离  不变）；编辑回包/启停 toggle/删除仅规则 owner 与 admin（PUT / DELETE 非 owner
      且非 admin → 403 forbidden，且 403 不泄露规则存在性——对非 admin 未知 ruleId 同样 403）；
      「Mock 此请求」创建新规则所有开发者可用，创建后自己即 owner。存量规则（owner 为空，v0.11.0
      之前创建）仅 admin 可管理，dev 只读 403。分享持久化：ShareSnapshot 随 data.json 落盘，
      重启不失效，7 天 TTL（expiresAt = 创建时刻 + 7 天）不变——过期链接返回 404，janitor 每小时
      清理过期快照（与  48h 会话流量保留并存）。
    - 容量与留存：会话流量上限契约化（--capture-session-max-entries，默认 20000，
      超限丢最旧 rolling window）；服务端过滤下推（GET /sessions/{id}/traffic 新增
      keyword/statusCode/from/to，过滤先于分页）；会话结束后流量保留 48h（--capture-traffic-retention-hours，
      默认 48，自结束时刻起算）供历史日志查询，之后由 janitor 每小时清理；历史日志查询期间保留期内
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /traffic:
    post:
      tags: [traffic]
      summary: 批量上传流量（SDK）
      description: |
        SDK 在抓包会话有效期间批量上传一次/多次请求的完整信息（全量抓包，会话内临时记录）。
        攒批：SDK 按条目数或累计字节阈值触发上传，单批 ≤ 500 条；
        服务端请求体上限 10MB（mockd Admin BodyLimit），SDK 应控制批次大小避免整批被拒。
        部分接受：entries 中字段不合法的条目被丢弃、合法条目正常入库，响应 count = 实际接受数；
        全部条目不合法返回 400。
        临时语义：会话流量按 MaxSessionTrafficEntries
        （--capture-session-max-entries，默认 20000）上限保留，超限丢最旧（rolling window）；
        会话结束（页面关闭 / 心跳超时 / 强制断开）后流量保留 48h（--capture-traffic-retention-hours，
        默认 48，自结束时刻起算）供历史日志查询，期满由 janitor 清理；
        命中 Mock 的请求仍按临时流量上传（mocked=true，其响应即 Mock 回包），随会话保留；
        持久化保留的是规则及其 source 快照（原始值+Mock 值），不是每条命中流量。
        仅接受激活会话内的流量：sessionId 不存在 → 404 session_not_found；会话已结束 → 409
        session_ended（保留期内不可再上传）。
      operationId: uploadTraffic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrafficUploadRequest'
      responses:
        '202':
          description: 已接收（含部分接受；count 为实际接受条目数）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrafficUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 会话已结束（session_ended）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token
      description: |
        两种凭证并存：
        - 会话 token：登录（POST /auth/login）签发、服务端持久化；Web 每次请求携带
          `Authorization: Bearer <token>`。有效期 7 天（服务端配置项）；登出即服务端吊销。
        - 长效 API Key：POST /auth/keys 创建，`Authorization: Bearer <mnpk_...>`；无过期，
          明文仅创建时返回一次，服务端只存 SHA-256 哈希（S3：任何日志不落 Authorization/key 明文）。
        两者权限等价（API Key 归属用户角色），401/403 语义一致。
...
  responses:
    BadRequest:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务端错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    TrafficEntry:
      type: object
      required: [timestamp, method, url, durationMs]
      properties:
        id:
          type: string
          description: 流量记录 ID（服务端生成）
        sessionId:
          type: string
          description: 所属会话 ID（服务端写入）
        timestamp:
          type: string
          format: date-time
          description: 请求发起时间（RFC3339，SDK 打点）
        method:
          type: string
          example: POST
        url:
          type: string
          description: 完整请求 URL
        path:
          type: string
          description: URL 路径（用于 Mock 匹配）
        query:
          type: string
          description: Query 字符串（不含 ?，原始顺序；无则空串）
        requestHeaders:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: 请求头（多值）
        requestBody:
          type: string
          description: |
            请求体：按 UTF-8 文本保存，超过 1MB 截断（ 固定上限，不做配置项）；
            无法 UTF-8 解码的二进制内容以 "[binary N bytes]" 占位保存。
        requestBodyBase64:
          type: string
          description: 二进制请求体的 base64（无法 UTF-8 解码时填充），供"Mock 此请求"回放原始字节
        requestBodyDecoded:
          type: string
          description: |
            请求体经 App 注入的协议解码器解出的可读 UTF-8 文本。
            仅作 Web 展示；存在时 Web 优先展示它，无则回退 requestBody 的 "[binary N bytes]" 占位。
            不参与回放；解码失败/未注入解码器时缺省。
        statusCode:
          type: integer
          nullable: true
          description: 响应状态码（请求失败时为空）
        responseHeaders:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          nullable: true
          description: 响应头（请求失败时为空）
        responseBody:
          type: string
          nullable: true
          description: |
            响应体：按 UTF-8 文本保存，超过 1MB 截断（ 固定上限，不做配置项）；
            无法 UTF-8 解码的二进制内容以 "[binary N bytes]" 占位保存；请求失败时为空。
        responseBodyBase64:
          type: string
          description: 二进制响应体的 base64（无法 UTF-8 解码时填充），供"Mock 此请求"回放原始字节；请求失败时为空
        responseBodyDecoded:
          type: string
          description: |
            响应体经 App 注入的协议解码器解出的可读 UTF-8 文本。
            仅作 Web 展示；存在时 Web 优先展示它，无则回退 responseBody 的 "[binary N bytes]" 占位。
            不参与回放——回放永远走 responseBodyBase64 原始字节；解码失败/未注入解码器时缺省。
        error:
          type: string
          nullable: true
          description: 错误信息（请求失败时；成功时为空）
        durationMs:
          type: integer
          description: 请求耗时（毫秒，请求发起至收到完成回调）
        mocked:
          type: boolean
          default: false
          description: 是否命中 Mock（ 恒为 false； 起标记，Web 请求流标注"Mock 命中"）
        matchedRuleId:
          type: string
          description: >-
            命中的 Mock 规则 ID（v0.13.0， P1-1）：仅 mocked=true 时非空，
            用于确认该请求究竟命中哪一条规则；真实流量为空/缺省。
        seq:
          type: integer
          description: |
            每会话自增的到达序号：服务端按上传到达顺序从 1 递增，
            供增量拉取（since）使用——ID 为 ULID，同毫秒内字典序不严格单调，故新增独立 seq。
            仅服务端与查询方（MCP/CLI/Web）消费，SDK 不感知、不消费。

    # -------------------------------------------------  机器通道（v0.12.0）
...
    TrafficUploadRequest:
      type: object
      required: [app, did, sessionId, entries]
      properties:
        app:
          type: string
          maxLength: 128
        did:
          type: string
          maxLength: 128
        sessionId:
          type: string
          description: 目标会话 ID（上传仅接受激活会话内流量）
        entries:
          type: array
          maxItems: 500
          items:
            $ref: '#/components/schemas/TrafficEntry'
          description: |
            一次批量上传的流量条目。SDK 攒批：按条目数或累计字节阈值触发上传，
            单批 ≤ 500 条且控制总字节，避免超过服务端 10MB 请求体上限。
    TrafficUploadResponse:
      type: object
      required: [accepted, count]
      properties:
        accepted:
          type: boolean
          description: 是否已接收（202 隐含接收， 恒为 true；false 仅作扩展预留）
        count:
          type: integer
          description: 实际接受的条目数（部分接受：合法条目数；全部条目不合法时服务端返回 400 而非 202）

    # ------------------------------------------------------------ Mock 规则
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: 机器可读错误码（如 device_not_registered）
        message:
          type: string
          description: 人类可读错误信息